+++ superwall/src/main/java/com/superwall/sdk/paywall/presentation/internal/operators/StorePresentationObjects.kt
@@ -19,5 +19,5 @@ internal suspend fun Superwall.storePresentationObjects(
         request,
         paywallStatePublisher
     )
-    presentationItems.setLast(lastPaywallPresentation)
+    presentationItems.last = lastPaywallPresentation
 }